+++ src/main/scala/edu/ie3/simona/agent/grid/GridAgentBuilder.scala
@@ -71,7 +71,7 @@ object GridAgentBuilder {
       thermalIslandGridsByBusId: Map[UUID, ThermalGrid],
   )(using
       constantData: GridAgentConstantData,
-      gridAgentContext: ActorContext[GridAgent.Request],
+      gridAgentContext: ActorContext[GridAgent.Message],
       log: Logger,
   ): Map[UUID, Set[ActorRef[ParticipantAgent.Request]]] = {
 
@@ -169,7 +169,7 @@ object GridAgentBuilder {
       thermalIslandGridsByBusId: Map[UUID, ThermalGrid],
   )(using
       constantData: GridAgentConstantData,
-      gridAgentContext: ActorContext[GridAgent.Request],
+      gridAgentContext: ActorContext[GridAgent.Message],
   ): Map[UUID, Set[ActorRef[ParticipantAgent.Request]]] = {
     participants
       .map { participant =>
@@ -217,7 +217,7 @@ object GridAgentBuilder {
       previousLevelEms: Map[UUID, ActorRef[FlexResponse]] = Map.empty,
   )(using
       constantData: GridAgentConstantData,
-      gridAgentContext: ActorContext[GridAgent.Request],
+      gridAgentContext: ActorContext[GridAgent.Message],
   ): Map[UUID, ActorRef[FlexResponse]] = {
     // For the current level, split controlled and uncontrolled EMs.
     // Uncontrolled EMs can be built right away.
@@ -281,7 +281,7 @@ object GridAgentBuilder {
       maybeControllingEm: Option[ActorRef[FlexResponse]],
   )(using
       constantData: GridAgentConstantData,
-      gridAgentContext: ActorContext[GridAgent.Request],
+      gridAgentContext: ActorContext[GridAgent.Message],
   ): ActorRef[ParticipantAgent.Request] = {
 
     val serviceMap: Map[ServiceType, ServiceRef] =
@@ -439,7 +439,7 @@ object GridAgentBuilder {
       scheduler: ActorRef[SchedulerMessage],
       maybeControllingEm: Option[ActorRef[FlexResponse]],
   )(using
-      gridAgentContext: ActorContext[GridAgent.Request]
+      gridAgentContext: ActorContext[GridAgent.Message]
   ): ActorRef[ParticipantAgent.Request] = {
 
     val key = ScheduleLock.singleKey(gridAgentContext, scheduler, PRE_INIT_TICK)
@@ -479,7 +479,7 @@ object GridAgentBuilder {
       maybeControllingEm: Option[ActorRef[FlexResponse]],
   )(using
       constantData: GridAgentConstantData,
-      gridAgentContext: ActorContext[GridAgent.Request],
+      gridAgentContext: ActorContext[GridAgent.Message],
   ): ActorRef[FlexResponse] =
     gridAgentContext.spawn(
       EmAgent(