+++ src/main/scala/edu/ie3/simona/agent/grid/GridAgentMessages.scala
@@ -77,23 +77,15 @@ object GridAgentMessages {
   final case class FinishGridSimulationTrigger(tick: Long)
       extends GridAgent.InternalRequest
 
-  /** Wrapper for activation values
-    *
-    * @param activation
-    *   the tick
-    */
-  final case class WrappedActivation(activation: Activation)
-      extends GridAgent.InternalRequest
-
   /** Trait for values that can be received as a response to a
-    * [[GridAgent.Request]].
+    * [[GridAgent.Message]].
     */
   sealed trait ReceivedValues extends GridAgent.InternalReply
 
   private type PowerRequestResponse[T] = (ActorRef[T], PowerResponse)
 
   private type SlackVoltageRequestResponse =
-    (ActorRef[GridAgent.Request], SlackVoltageResponse)
+    (ActorRef[GridAgent.Message], SlackVoltageResponse)
 
   sealed trait ReceivedPowerValues extends ReceivedValues {
     def values: Vector[(ActorRef[_], PowerResponse)]
@@ -114,7 +106,7 @@ object GridAgentMessages {
     *   the grid power values and their senders
     */
   final case class ReceivedGridPowerValues(
-      values: Vector[PowerRequestResponse[GridAgent.Request]]
+      values: Vector[PowerRequestResponse[GridAgent.Message]]
   ) extends ReceivedPowerValues
 
   /** Wrapper for received slack voltage values (v)
@@ -146,7 +138,7 @@ object GridAgentMessages {
   final case class RequestGridPower(
       currentSweepNo: Int,
       nodeUuids: Seq[UUID],
-      sender: ActorRef[GridAgent.Request],
+      sender: ActorRef[GridAgent.Message],
   ) extends GridAgent.InternalRequest
 
   sealed trait PowerResponse extends GridAgent.InternalReply
@@ -212,7 +204,7 @@ object GridAgentMessages {
   final case class SlackVoltageRequest(
       currentSweepNo: Int,
       nodeUuids: Seq[UUID],
-      sender: ActorRef[GridAgent.Request],
+      sender: ActorRef[GridAgent.Message],
   ) extends GridAgent.InternalRequest
 
   /** Provide complex voltage at the nodes that the sender's sub grid shares