+++ src/main/scala/edu/ie3/simona/sim/setup/SimonaStandaloneSetup.scala
@@ -65,7 +65,7 @@ class SimonaStandaloneSetup(
       context: ActorContext[_],
       environmentRefs: EnvironmentRefs,
       resultEventListeners: Seq[ActorRef[ResultEvent]],
-  ): Iterable[ActorRef[GridAgent.Request]] = {
+  ): Iterable[ActorRef[GridAgent.Message]] = {
 
     /* get the grid */
     val subGridTopologyGraph = GridProvider
@@ -297,7 +297,7 @@ class SimonaStandaloneSetup(
       context: ActorContext[_],
       environmentRefs: EnvironmentRefs,
       resultEventListeners: Seq[ActorRef[ResultEvent]],
-  ): Map[Int, ActorRef[GridAgent.Request]] = {
+  ): Map[Int, ActorRef[GridAgent.Message]] = {
     subGridTopologyGraph
       .vertexSet()
       .asScala