+++ src/test/scala/edu/ie3/simona/agent/grid/DBFSAlgorithmFailedPowerFlowSpec.scala
@@ -82,7 +82,7 @@ class DBFSAlgorithmFailedPowerFlowSpec
 
     // since the grid agent is stopped after a failed power flow
     // we need to initialize the agent for each test
-    def initAndGoToSimulateGrid: ActorRef[GridAgent.Request] = {
+    def initAndGoToSimulateGrid: ActorRef[GridAgent.Message] = {
       val centerGridAgent =
         testKit.spawn(
           GridAgent(
@@ -128,7 +128,7 @@ class DBFSAlgorithmFailedPowerFlowSpec
       scheduleActivationMsg.unlockKey shouldBe Some(key)
 
       // send init data to agent
-      centerGridAgent ! WrappedActivation(Activation(3600))
+      centerGridAgent ! Activation(3600)
 
       // we expect a completion message
       scheduler.expectMessageType[Completion].newTick shouldBe Some(3600)
@@ -142,8 +142,7 @@ class DBFSAlgorithmFailedPowerFlowSpec
       val sweepNo = 0
 
       // send the start grid simulation trigger
-      centerGridAgent ! WrappedActivation(Activation(3600))
-
+      centerGridAgent ! Activation(3600)
       // we expect a request for grid power values here for sweepNo $sweepNo
       val powerRequestSender = inferiorGridAgent.expectGridPowerRequest()
 
@@ -226,7 +225,7 @@ class DBFSAlgorithmFailedPowerFlowSpec
       val sweepNo = 0
 
       // send the start grid simulation trigger
-      centerGridAgent ! WrappedActivation(Activation(3600))
+      centerGridAgent ! Activation(3600)
 
       // we expect a request for grid power values here for sweepNo 0
       val powerRequestSender = inferiorGridAgent.expectGridPowerRequest()
@@ -299,7 +298,7 @@ class DBFSAlgorithmFailedPowerFlowSpec
       val hvGridAgent =
         InferiorGA(TestProbe("HvGridAgent"), Seq(supNodeA.getUuid))
 
-      val slackGridAgent: ActorRef[GridAgent.Request] = testKit.spawn(
+      val slackGridAgent: ActorRef[GridAgent.Message] = testKit.spawn(
         GridAgent(
           environmentRefs,
           simonaConfig, // stopOnFailure is enabled
@@ -334,13 +333,13 @@ class DBFSAlgorithmFailedPowerFlowSpec
       scheduleActivationMsg.unlockKey shouldBe Some(key)
 
       // send init data to agent
-      slackGridAgent ! WrappedActivation(Activation(3600))
+      slackGridAgent ! Activation(3600)
 
       // we expect a completion message
       scheduler.expectMessageType[Completion].newTick shouldBe Some(3600)
 
       // send the start grid simulation trigger
-      slackGridAgent ! WrappedActivation(Activation(3600))
+      slackGridAgent ! Activation(3600)
 
       val powerRequestSender = hvGridAgent.expectGridPowerRequest()