+++ src/test/scala/edu/ie3/simona/agent/grid/DBFSAlgorithmParticipantSpec.scala
@@ -87,7 +87,7 @@ class DBFSAlgorithmParticipantSpec
     s"initialize itself when it receives an init activation" in {
 
       // this subnet has 1 superior grid (ehv) and 3 inferior grids (mv). Map the gates to test probes accordingly
-      val subGridGateToActorRef: Map[SubGridGate, ActorRef[GridAgent.Request]] =
+      val subGridGateToActorRef: Map[SubGridGate, ActorRef[GridAgent.Message]] =
         hvSubGridGates.map { gate =>
           gate -> superiorGridAgent.ref
         }.toMap
@@ -148,7 +148,7 @@ class DBFSAlgorithmParticipantSpec
     s"go to SimulateGrid when it receives an activity start trigger" in {
 
       // send init data to agent
-      gridAgentWithParticipants ! WrappedActivation(Activation(3600))
+      gridAgentWithParticipants ! Activation(3600)
 
       // we expect a completion message
       scheduler.expectMessageType[Completion].newTick shouldBe Some(3600)
@@ -160,7 +160,7 @@ class DBFSAlgorithmParticipantSpec
 
       // send the start grid simulation trigger
       // the gird agent should send a RequestAssetPowerMessage to the load agent
-      gridAgentWithParticipants ! WrappedActivation(Activation(3600))
+      gridAgentWithParticipants ! Activation(3600)
 
       // we expect a request for voltage values of our slack node
       // (voltages are requested by our agent under test from the superior grid)