+++ src/test/scala/edu/ie3/simona/agent/grid/ThermalGridIT.scala
@@ -112,7 +112,7 @@ class ThermalGridIT
         simulationEnd = simulationEndDate,
       )
 
-      val gridAgent = TestProbe[GridAgent.Request]("GridAgent")
+      val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
       val resultListener = TestProbe[ResultEvent]("ResultListener")
       val scheduler: TestProbe[SchedulerMessage] = TestProbe("scheduler")
       val primaryServiceProxy =
@@ -736,7 +736,7 @@ class ThermalGridIT
         simulationEnd = simulationEndWithPv,
       )
 
-      val gridAgent = TestProbe[GridAgent.Request]("GridAgent")
+      val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
       val resultListener: TestProbe[ResultEvent] = TestProbe("resultListener")
       val scheduler: TestProbe[SchedulerMessage] = TestProbe("scheduler")
       val primaryServiceProxy =