+++ src/test/scala/edu/ie3/simona/agent/participant/ParticipantAgentSpec.scala
@@ -91,7 +91,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
       "calculate operating point and results correctly with no additional model activations" in {
 
         val scheduler = createTestProbe[SchedulerMessage]()
-        val gridAgent = createTestProbe[GridAgent.Request]()
+        val gridAgent = createTestProbe[GridAgent.Message]()
         val resultListener = createTestProbe[ResultEvent]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
 
@@ -247,7 +247,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
       "calculate operating point and results correctly with additional model activations" in {
 
         val scheduler = createTestProbe[SchedulerMessage]()
-        val gridAgent = createTestProbe[GridAgent.Request]()
+        val gridAgent = createTestProbe[GridAgent.Message]()
         val resultListener = createTestProbe[ResultEvent]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
 
@@ -401,7 +401,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
       "calculate operating point and results correctly with additional model activations" in {
 
         val scheduler = createTestProbe[SchedulerMessage]()
-        val gridAgent = createTestProbe[GridAgent.Request]()
+        val gridAgent = createTestProbe[GridAgent.Message]()
         val resultListener = createTestProbe[ResultEvent]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
         val service = createTestProbe[ServiceMessage]()
@@ -669,7 +669,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
       "calculate operating point and results correctly" in {
 
         val scheduler = createTestProbe[SchedulerMessage]()
-        val gridAgent = createTestProbe[GridAgent.Request]()
+        val gridAgent = createTestProbe[GridAgent.Message]()
         val resultListener = createTestProbe[ResultEvent]()
         val service = createTestProbe[ServiceMessage]()
 
@@ -891,7 +891,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
       "calculate operating point and results correctly with no additional model activations" in {
 
         val em = createTestProbe[FlexResponse]()
-        val gridAgent = createTestProbe[GridAgent.Request]()
+        val gridAgent = createTestProbe[GridAgent.Message]()
         val resultListener = createTestProbe[ResultEvent]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
 
@@ -1070,7 +1070,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
       "calculate operating point and results correctly with additional model activations" in {
 
         val em = createTestProbe[FlexResponse]()
-        val gridAgent = createTestProbe[GridAgent.Request]()
+        val gridAgent = createTestProbe[GridAgent.Message]()
         val resultListener = createTestProbe[ResultEvent]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
 
@@ -1308,7 +1308,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
       "calculate operating point and results correctly with additional model activations" in {
 
         val em = createTestProbe[FlexResponse]()
-        val gridAgent = createTestProbe[GridAgent.Request]()
+        val gridAgent = createTestProbe[GridAgent.Message]()
         val resultListener = createTestProbe[ResultEvent]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
         val service = createTestProbe[ServiceMessage]()
@@ -1713,7 +1713,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
       "calculate operating point and results correctly" in {
 
         val em = createTestProbe[FlexResponse]()
-        val gridAgent = createTestProbe[GridAgent.Request]()
+        val gridAgent = createTestProbe[GridAgent.Message]()
         val resultListener = createTestProbe[ResultEvent]()
         val service = createTestProbe[ServiceMessage]()