+++ src/test/scala/edu/ie3/simona/model/participant/evcs/EvcsModelIT.scala
@@ -92,7 +92,7 @@ class EvcsModelIT
 
     "handle a few requests and arrivals as expected" in {
 
-      val gridAgent = TestProbe[GridAgent.Request]("GridAgent")
+      val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
       val resultListener = TestProbe[ResultEvent]("ResultListener")
       val primaryServiceProxy =
         TestProbe[PrimaryServiceProxy.Message]("PrimaryServiceProxy")